requirements.txt
django==4.2.16
djangorestframework==3.15.2
django-environ==0.11.2
django-cors-headers==4.4.0
channels==4.1.0
channels_redis==4.2.0
redis==5.0.8
celery==5.4.0
openai==1.46.0
PyPDF2==3.0.1
python-magic==0.4.27
psycopg2==2.9.9  # Changed from psycopg2-binary
djangorestframework-simplejwt==5.3.1
django-celery-beat==2.7.0